Multilingual_Invoice_extraction_app/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
pandas 
openpyxl 
python-docx